requirements.txt
django==1.10.5
psycopg2==2.6.2
py-bcrypt==0.4
flake8==3.2.1
